Adds scroll to bottom functionality for chat
Adds a "scroll to bottom" button to the chat interface.
This allows users to easily scroll to the bottom of the chat history,
especially when new messages are added or when the user is not at the bottom of the chat.
Implements an IntersectionObserver to detect if the user is at the bottom.

Signed-off-by: TheGloved1 <gloves1229@gmail.com>